feat: 新增评议单管理相关功能

+++ ruoyi-common/src/main/java/com/ruoyi/common/core/controller/BaseController.java
@@ -86,7 +86,7 @@ public class BaseController
         rspData.setCode(HttpStatus.SUCCESS);
         rspData.setMsg("查询成功");
         rspData.setRows(list);
-        rspData.setTotal(new PageInfo(list).getTotal());
+        rspData.setTotal((int) new PageInfo(list).getTotal());
         return rspData;
     }
